% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FloodComplex.R
\name{as_filtration}
\alias{as_filtration}
\title{Convert a flood_complex object to a filtration list}
\usage{
as_filtration(fc)
}
\arguments{
\item{fc}{A \code{"flood_complex"} object.}
}
\value{
A filtration list compatible with \code{boundary_info}.
}
\description{
Convert a flood_complex object to a filtration list
}
