% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Betti.R
\name{betti_number}
\alias{betti_number}
\title{Safely compute the rank of a sparse matrix}
\usage{
betti_number(simplices, bound_dim, eps = NULL)
}
\arguments{
\item{eps}{Optional numerical tolerance to pass to \code{rankMatrix()}.}

\item{mat}{A matrix (usually sparse) whose rank is to be computed.}
}
\value{
An integer representing the rank of the matrix.
}
\description{
This helper function wraps \code{Matrix::rankMatrix()} to safely handle empty matrices (i.e., with 0 rows or columns).
}
\keyword{internal}
