% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filtration.R
\name{build_vr_filtration}
\alias{build_vr_filtration}
\title{Vietoris-Rips Filtration}
\usage{
build_vr_filtration(points, eps_max)
}
\arguments{
\item{points}{Data point input}

\item{eps_max}{Maximum scale (epsilon)}
}
\description{
Get the boundary matrix and its reduction information in matrix form
}
