% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlphaComplex.R
\name{DelaunayComplex}
\alias{DelaunayComplex}
\title{Construct a Delaunay Complex}
\usage{
DelaunayComplex(points)
}
\arguments{
\item{points}{A numeric matrix or data.frame with one point per row
(columns are coordinates); must be in "general position".}
}
\value{
A list of class \code{c("delaunay_complex", "alpha_complex")};
  see \code{\link{AlphaComplex}} for the element description.
}
\description{
Construct a Delaunay Complex
}
\examples{
points <- matrix(c(0, 1, 1, 0, 0, 0, 1, 1), ncol = 2)
delaunay_complex <- DelaunayComplex(points)
}
