% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPD.R
\name{plot_persistence}
\alias{plot_persistence}
\title{Plot Persistence Diagram}
\usage{
plot_persistence(df)
}
\arguments{
\item{df}{Dataframe from plot_persistence.}
}
\value{
A ggplot2 object representing the persistence diagram.
}
\description{
Plot Persistence Diagram
}
\examples{
points <- matrix(c(0, 1, 1, 0, 0, 0, 1, 1), ncol = 2)
filtration <- build_filtration(points, method = "VR", eps_max = 1.2)
res <- boundary_info(filtration)
pairs <- extract_persistence_pairs(filtration, res$last_1, res$pivot_owner)
plot_persistence(pairs)
}
