% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filtration.R
\name{vr_scale_of_simplex}
\alias{vr_scale_of_simplex}
\title{Distance scale of a simplex}
\usage{
vr_scale_of_simplex(points, simplex_idx)
}
\arguments{
\item{points}{Data point input}

\item{simplex_idx}{Index of points in the simplex}
}
\description{
Get the boundary matrix and its reduction information in matrix form
}
