% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Abstractsimplicialcomplex.R
\name{abstract_simplicial_complex}
\alias{abstract_simplicial_complex}
\title{Compute Euler characteristic for an abstract simplicial complex}
\usage{
abstract_simplicial_complex(simplices, dimension, tol = NULL)
}
\arguments{
\item{simplices}{A list of simplices (each a numeric vector).}

\item{dimension}{Optional max dimension to compute up to.}

\item{tol}{Optional numerical tolerance to pass to \code{rankMatrix()}.}
}
\value{
The Euler characteristic \eqn{\chi}.
}
\description{
Compute Euler characteristic for an abstract simplicial complex
}
