% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Persistence.R
\name{boundary_info}
\alias{boundary_info}
\title{Get the boundary matrix and its reduction information in matrix form}
\usage{
boundary_info(F)
}
\arguments{
\item{F}{Filtration list, each element includes simplex and time}
}
\description{
Get the boundary matrix and its reduction information in matrix form
}
