% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Persistence.R
\name{extract_persistence_pairs}
\alias{extract_persistence_pairs}
\title{This function extracts the persistence from combining the
boundary matrix and its filtration}
\usage{
extract_persistence_pairs(F, last_1, pivot_owner)
}
\arguments{
\item{F}{Filtration list, each element includes simplex and time}

\item{last_1}{The last 1 row index for each column in boundary matrix}

\item{pivot_owner}{The column index owning the pivot row}
}
\description{
This function extracts the persistence from combining the
boundary matrix and its filtration
}
